+++ casbah-core/src/main/scala/MongoCollection.scala
@@ -193,6 +193,7 @@ trait MongoCollectionBase extends Logging { self =>
    * @param options - see Bytes QUERYOPTION_*
    * @return the objects, if found
    */
+  @deprecated("Use `find().skip().batchSize()`.", "2.7")
   def find[A <% DBObject, B <% DBObject](ref: A, fields: B, numToSkip: Int, batchSize: Int) =
     _newCursor(underlying.find(ref, fields, numToSkip, batchSize))
 
@@ -608,7 +609,7 @@ trait MongoCollectionBase extends Logging { self =>
    * @param q search query for old object to update
    * @param o object with which to update <tt>q</tt>
    */
-  @deprecated("In the face of default arguments this is a bit silly. Please use update(multi=True)", "2.3.0")
+  @deprecated("In the face of default arguments this is a bit silly. Please use `update(multi=True)`.", "2.3.0")
   def updateMulti[A <% DBObject, B <% DBObject](q: A, o: B) = underlying.updateMulti(q, o)
 
   override def hashCode() = underlying.hashCode
@@ -781,9 +782,8 @@ trait MongoCollectionBase extends Logging { self =>
 
   /**
    * Sets queries to be OK to run on slave nodes.
-   * @deprecated Replaced with ReadPreference.SECONDARY
    */
-  @deprecated("Replaced with ReadPreference.SECONDARY", "2.3.0")
+  @deprecated("Replaced with `ReadPreference.SECONDARY`", "2.3.0")
   def slaveOk() = underlying.slaveOk() // use parens because this side-effects
 
   /**