+++ casbah-core/src/main/scala/MongoOptions.scala
@@ -35,7 +35,9 @@ import javax.net.SocketFactory
  *
  * @since 2.0
  * @see com.mongodb.MongoOptions
+ *
  */
+@deprecated("Please use MongoClient and MongoClientOptions", "2.7")
 object MongoOptions {
 
   val Defaults = new MongoOptions()